apps/02-personal-md-files/backend/python/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2 